Rename ErrorHandler namespace to ErrorHandling
