+++ src/DI/Plugin/CoreServicesPlugin.php
@@ -7,11 +7,11 @@ use Apitte\Core\Application\IApplication;
 use Apitte\Core\Dispatcher\IDispatcher;
 use Apitte\Core\Dispatcher\JsonDispatcher;
 use Apitte\Core\Dispatcher\WrappedDispatcher;
-use Apitte\Core\ErrorHandler\ErrorConverter;
-use Apitte\Core\ErrorHandler\IErrorHandler;
-use Apitte\Core\ErrorHandler\JsonErrorConverter;
-use Apitte\Core\ErrorHandler\PsrLogErrorHandler;
-use Apitte\Core\ErrorHandler\SimpleErrorHandler;
+use Apitte\Core\ErrorHandling\ErrorConverter;
+use Apitte\Core\ErrorHandling\IErrorHandler;
+use Apitte\Core\ErrorHandling\JsonErrorConverter;
+use Apitte\Core\ErrorHandling\PsrLogErrorHandler;
+use Apitte\Core\ErrorHandling\SimpleErrorHandler;
 use Apitte\Core\Handler\IHandler;
 use Apitte\Core\Handler\ServiceHandler;
 use Apitte\Core\Router\IRouter;
