+++ tests/cases/Dispatcher/WrappedDispatcher.phpt
@@ -7,8 +7,8 @@
 require_once __DIR__ . '/../../bootstrap.php';
 
 use Apitte\Core\Dispatcher\WrappedDispatcher;
-use Apitte\Core\ErrorHandler\JsonErrorConverter;
-use Apitte\Core\ErrorHandler\SimpleErrorHandler;
+use Apitte\Core\ErrorHandling\JsonErrorConverter;
+use Apitte\Core\ErrorHandling\SimpleErrorHandler;
 use Apitte\Core\Http\ApiRequest;
 use Apitte\Core\Http\ApiResponse;
 use Contributte\Psr7\Psr7ResponseFactory;
